requirements.txt
absl-py==2.3.0
aiofiles==24.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyascii==0.3.2
anyio==4.9.0
audioread==3.0.1
av==14.4.0
azure-cognitiveservices-speech==1.44.0
azure-core==1.34.0
babel==2.17.0
beautifulsoup4==4.13.2
bitarray==3.4.2
blinker==1.9.0
blis==1.2.1
boto3==1.38.39
botocore==1.38.39
Brotli==1.1.0
cached_path==1.7.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cn2an==0.5.22
colorama==0.4.6
comtypes==1.4.11
confection==0.1.5
cymem==2.0.11
Cython==3.1.2
dateparser==1.1.8
decorator==5.2.1
Deprecated==1.2.18
Distance==0.1.3
distro==1.9.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
eng_to_ipa==0.0.2
exceptiongroup==1.3.0
fairseq==0.12.2
faiss-cpu==1.11.0
fastapi==0.115.13
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.6.0
filelock==3.18.0
fish-audio-sdk==2025.6.3
Flask==3.1.1
fsspec==2025.5.1
fugashi==1.3.0
future==1.0.0
g2p-en==2.1.0
g2pkk==0.1.2
gmpy2==2.2.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
gradio==5.34.1
gradio_client==1.10.3
groovy==0.1.2
grpcio==1.73.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface-hub==0.33.0
hydra-core==1.0.7
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
inflect==7.0.0
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonlines==1.2.0
langcodes==3.5.0
langid==1.1.6
language_data==1.3.0
lazy_loader==0.4
librosa==0.9.1
llvmlite==0.44.0
loguru==0.7.2
lxml==5.4.0
marisa-trie==1.2.1
Markdown==3.8.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mecab-python3==1.0.9
mkl-service==2.4.0
mkl_fft==1.3.14
mkl_random==1.2.10
mpmath==1.3.0
msgpack==1.1.1
murmurhash==1.0.13
networkx==2.8.8
nltk==3.9.1
num2words==0.5.12
numba==0.61.2
numpy==1.26.4
omegaconf==2.0.6
openai==1.88.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.0
pastel==0.2.1
pillow==11.2.1
plac==1.4.5
platformdirs==4.3.8
poethepoet==0.24.4
pooch==1.8.2
portalocker==3.2.0
praat-parselmouth==0.4.6
preshed==3.0.10
priority==2.0.0
proces==0.1.7
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pykakasi==2.2.1
pypinyin==0.50.0
PySocks==1.7.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyttsx3==2.98
pytz==2025.2
pyworld==0.3.5
PyYAML==6.0.2
Quart==0.20.0
regex==2024.11.6
requests==2.32.4
resampy==0.4.3
rich==13.9.4
rsa==4.9.1
ruff==0.12.0
s3transfer==0.13.0
sacrebleu==2.5.1
safehttpx==0.1.6
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlite-vec==0.1.6
srsly==2.5.1
sse-starlette==2.3.6
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
taskgroup==0.2.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.3
torch==2.5.1
torchaudio==2.5.1
torchcrepe==0.0.22
torchvision==0.20.1
tqdm==4.67.1
transformers==4.52.4
txtsplit==1.0.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.3.7
unidic==1.1.0
unidic-lite==1.0.8
urllib3==2.5.0
uvicorn==0.34.3
wasabi==0.10.1
Wave==0.0.2
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
win-inet-pton==1.1.0
win32_setctime==1.2.0
wrapt==1.17.2

wsproto==1.2.0
google-genai==1.0.0
edge-tts==7.2.8
